requirements.txt
Django
psycopg2
django-import-export